backend/requirements.txt
Django==3.2
djongo==1.3.6
pymongo
dnspython
python-dotenv
PyJWT
django-cors-headers
